% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portkey.R
\name{flipped.portkey}
\alias{flipped.portkey}
\title{Flipped portkey two-coin algorithm: returns the result of the accept-reject step of a Flipped Portkey Barker's acceptance probability.
Note: the flipped portkey is the same as Barker's for \code{beta = 1}.}
\usage{
flipped.portkey(prop, curr, beta = 0.99, pf, Cprop, Ccurr, ...)
}
\arguments{
\item{prop}{The proposed value in the MCMC step}

\item{curr}{The current value in the MCMC step}

\item{beta}{The portkey parameter with default being .99. If \code{beta = 1}, this is the regular two-coin algorithm}

\item{pf}{A function that produces a Bern(p) realization, with argument \code{value} for the state 
of the Markov chain}

\item{Cprop}{Upper bound for the inverse target at the proposed value}

\item{Ccurr}{Upper bound for the inverse target at the current value}

\item{...}{additional arguments that go into \code{pf}}
}
\value{
a variable \code{x} which is either \code{curr} or \code{prop} and integer \code{loops} that returns the number 
    of loops the Bernoulli factory took
}
\description{
Flipped portkey two-coin algorithm: returns the result of the accept-reject step of a Flipped Portkey Barker's acceptance probability.
Note: the flipped portkey is the same as Barker's for \code{beta = 1}.
}
\references{
Vats, D., Gonçalves, F. B., Łatuszyński, K., Roberts G. O., Efficient Bernoulli factory MCMC for intractable posteriors (2020+), arxiv.
}
\author{
Dootika Vats, \email{dootika@iitk.ac.in}
}
