% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portkey.R
\name{twocoin}
\alias{twocoin}
\title{Two-coin algorithm}
\usage{
twocoin(prop, curr, pf, Cprop, Ccurr, ...)
}
\arguments{
\item{prop}{The proposed value in the MCMC step}

\item{curr}{The current value in the MCMC step}

\item{pf}{A function that produces a Bern(p) realization, with argument \code{value} for the state 
of the Markov chain}

\item{Cprop}{Upper bound for the proposed value}

\item{Ccurr}{Upper bound for the current value}

\item{...}{additional arguments that go into \code{pf}}
}
\value{
a variable \code{x} which is either \code{curr} or \code{prop} and integer \code{loops} that returns the number 
    of loops the Bernoulli factory took
}
\description{
Two-coin algorithm
}
